Create Sum_of_Consecutive_Powers
This problem was tricky. Given a range, it asked to return all the numbers in that range where the number (N) was equal too the
sum of each individual digit raised to the consecutive power. For example: 89 = 8^1 + 9^2. Also, 135 = 1^1 + 3^2 + 5^3. The goal was
to find the numbers that satisfied that quality given a certain range. Good learning tool for those who haven't used imap before.