Create World_Bits_War
This one was difficult for me because I'm still trying to wrap my head around counting in bits in the first place. The idea here is that, given a list of integers, the odds and the evens are at war. Each number adds a point to their side given how many '1's are in their binary equivalent. so 5 = 101 so that's 2 points for the odd side. The tricky thing about this particular one is that negative numbers come into play. Every negative number subtracts their '1's from their side's overall score. So [5,-3,2] would have 5 scoring two points for odds, -3 subtracting two points from odd, and 2 earning one point for evens so evens win. 

Lots of fun with this one, I decided to create a separate function just to count the number of 1s in the binary representation of any given integer. Then I just recorded the tally based on whether a number was odd/even and positive/negative and boom, to the victor goes the spoils.